ai-analytics-service/requirements.txt
fastapi==0.70.0
uvicorn[standard]==0.15.0
pandas==1.3.4
numpy==1.21.2
scikit-learn==0.24.2
tensorflow==2.6.0
torch==1.9.1
kafka-python==2.0.2
pyspark==3.1.2
redis==3.5.3
mlflow==1.20.2
python-dotenv==0.19.1
loguru==0.5.3
